Make node traversal APIs more type-safe
The callback that was previously being passed to `AANode::traverse` had
to operate in two entirely distinct modes, one for when it was being
called on the way down and one for going back up. Mixing the two
together led to an extra match inside everything passed as a callback
and meant the implementation had to deal with extra cases that could
never happen. Separating it into two callbacks allows all of that to be
cleaned up, meaning that it is no longer possible for the callback(s) to
return an inappropriate value and that we don't need a dependency on the
`log` crate for the bad case.

`traverse_mut` also gets a limited version of the same change applied;
since it doesn't call any callbacks on the way up, there is no second
callback to add, but we can statically ensure that a direction is never
returned at the top level.

The benchmarks showed regressions of 10-15% on the Contains runs but
similarly sized gains on the Remove runs.